SocketServer & SocketClient
Simple Server and Client program using sockets.
Run the server first. This creates a socket on port 4444.
Then run the client, it creates a socket and connects to the server socket.
Client passes a string in format "<0-6> , <string>"
Server parses the string and returns color at specified index in colors array, followed by the word.